Geeks For Geeks 7
A number N is given. Divide it in half by using the binary shift operator. Use the floor value if the result is in decimal. If the number is 1, leave it as it is.

Input: N = 1
Output: 1

Input: N = 37
Output: 18
Explaination: 37/2 = 18.5 which is in 
decimal form and whose floor value is 18.