Geeks for Geeks 19
Given a number 32 bit number N .Find the number you get by inverting bits in its binary representation (i.e. 1 is made 0 and 0 is made 1). In other words,Negate(~) the number.

Input:
N=4289384
Output:
4290677911

Explanation:
00000000010000010111001101101000
is the binary representation of 4289384.
Negating this in 32 bits, we get
11111111101111101000110010010111
which is the binary representation of
4290677911.

Input:
N=1
Output:
4294967294

Explanation:
Binary representation of 1 is
00000000000000000000000000000001.
Negating this we get
11111111111111111111111111111110
which is the binary representation of
4294967294.