Geeks for Geeks 21
Given a number, the task is to set all odd bits of a number.
NOTE: Position of Least Significant Bit is considered as 1.

Input: n = 20
Output: 21 

Explanation: Binary representation of 20 
is 10100. Setting all odd bits make the 
number 10101 which is binary
representation of 21.

Input: n = 10
Output: 15

Explanation: Binary representation of 10
is 1010. Setting all odd bits make the
number 1111 which is binary representation
of 15.