Geeks for Geeks 27
You are given an array A of size N. You need to do the following:

1.You need to xor the elements of the arrays with their respective indices(index starting from 0).
2.After step1, print the array.
3.Now, set all the elements of A to zero.
4.Now, print A.


Input Format:
The first line contains T, the number of testcases. T testcases follow. Each testcase contains two lines of input. The first line contains N, the size of the array. The second line contains the elements of the array.

Output Format:
For each testcase, in a new line, print the output.

Your Task:
Since this is a function problem, you don't need to take any input. Just complete the provided functions.

Example:
Input:
1
10
1 2 3 4 5 6 7 8 9 10

Output:
1 3 1 7 1 3 1 15 1 3
0 0 0 0 0 0 0 0 0 0